correcting some minor graphical error over text/value entry

+++ jgl/includes/widget_component/jgl_widget_component.h
@@ -206,7 +206,7 @@ public:
 	void 		set_value(float p_value){_value = p_value;calc_text();}
 	void 		resize(Vector2 p_anchor, Vector2 p_area)
 		{set_anchor(p_anchor);set_area(p_area);calc_text_size_height(_area);}
-	void 		calc_value(){ _value = atof(_text.c_str());}
+	void 		calc_value(){ _value = static_cast<float>(atof(_text.c_str()));}
 	void 		calc_text(){ _text = ftoa(_value);}
 
 		//Getter
